[#1375] Revise public API
move UsersResource and Factory to internal package
