+++ bundles/org.eclipse.passage.lic.users.model/META-INF/MANIFEST.MF
@@ -13,8 +13,8 @@ Require-Bundle: org.eclipse.core.runtime;bundle-version="0.0.0";x-installation:=
  org.eclipse.emf.ecore.xmi;bundle-version="0.0.0";visibility:=reexport,
  org.eclipse.passage.lic.emf;bundle-version="0.0.0",
  org.eclipse.passage.lic.users;bundle-version="0.0.0";visibility:=reexport
-Export-Package: org.eclipse.passage.lic.internal.users.model;
-  x-friends:="org.eclipse.passage.loc.dashboard.ui",
+Export-Package: org.eclipse.passage.lic.internal.users.model;x-friends:="org.eclipse.passage.loc.dashboard.ui",
+ org.eclipse.passage.lic.internal.users.model.util;x-internal:=true,
  org.eclipse.passage.lic.users.model.api,
  org.eclipse.passage.lic.users.model.impl;x-internal:=true,
  org.eclipse.passage.lic.users.model.meta,
