+++ bundles/org.eclipse.passage.lic.users.model/src/org/eclipse/passage/lic/internal/users/model/migration/UsersResourceHandler.java
@@ -26,6 +26,7 @@ import org.eclipse.passage.lic.users.model.api.User;
 import org.eclipse.passage.lic.users.model.api.UserOrigin;
 import org.eclipse.passage.lic.users.model.meta.UsersPackage;
 
+@SuppressWarnings("restriction")
 public final class UsersResourceHandler extends MigratingResourceHandler {
 
 	@Override
