+++ bundles/org.eclipse.passage.lic.users.model/src/org/eclipse/passage/lic/internal/users/model/util/UsersResourceImpl.java
@@ -1,0 +1,39 @@
+/*******************************************************************************
+ * Copyright (c) 2018, 2024 ArSysOp
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * https://www.eclipse.org/legal/epl-2.0/.
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *     ArSysOp - initial API and implementation
+ *******************************************************************************/
+package org.eclipse.passage.lic.internal.users.model.util;
+
+import java.util.Map;
+
+import org.eclipse.emf.common.util.URI;
+import org.eclipse.emf.ecore.xmi.impl.XMIResourceImpl;
+import org.eclipse.passage.lic.internal.users.model.migration.UsersResourceHandler;
+
+public class UsersResourceImpl extends XMIResourceImpl {
+
+	public UsersResourceImpl(URI uri) {
+		super(uri);
+	}
+
+	public UsersResourceImpl() {
+		super();
+	}
+
+	@Override
+	protected void init() {
+		super.init();
+		Map<Object, Object> options = getDefaultLoadOptions();
+		options.put(OPTION_RECORD_UNKNOWN_FEATURE, Boolean.TRUE);
+		options.put(OPTION_RESOURCE_HANDLER, new UsersResourceHandler());
+	}
+
+}
